// libs/admin-plugins/shipping/international-shipping/src/ui/generated-types.ts
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
  Date: any;
};
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type ShippingCountryPrice = Node & {
  __typename?: 'ShippingCountryPrice';
  id: Scalars['ID'];
  country_code: Scalars['String'];
  country_name: Scalars['String'];
  price: Scalars['Float'];
};

export type ShippingCountryPriceList = PaginatedList & {
  __typename?: 'ShippingCountryPriceList';
  items: Array<ShippingCountryPrice>;
  totalItems: Scalars['Int'];
};

export type CreateShippingCountryInput = {
  country_code: Scalars['String'];
  price: Scalars['Float'];
};

export type UpdateShippingCountryInput = {
  id: Scalars['ID'];
  price: Scalars['Float'];
};
// ...
  export type Variables = ShippingCountryQueryVariables;
// ...
export namespace CreateShippingCountry {
  export type Variables = CreateShippingCountryMutationVariables;
  export type Mutation = CreateShippingCountryMutation;
  export type CreateShippingCountry = (NonNullable<CreateShippingCountryMutation['createShippingCountry']>);
}

export namespace UpdateShippingCountry {
  export type Variables = UpdateShippingCountryMutationVariables;
  export type Mutation = UpdateShippingCountryMutation;
  export type UpdateShippingCountry = (NonNullable<UpdateShippingCountryMutation['updateShippingCountry']>);
}
// ...
export type ShippingCountryQueryVariables = Exact<{
  id: Scalars['ID'];
}>;


export type ShippingCountryQuery = (
  { __typename?: 'Query' }
  & { shippingCountry?: Maybe<(
    { __typename?: 'ShippingCountryPrice' }
    & Pick<ShippingCountryPrice, 'id' | 'price' | 'country_code' | 'country_name'>
  )> }
);

export type CreateShippingCountryMutationVariables = Exact<{
  input: CreateShippingCountryInput;
}>;


export type CreateShippingCountryMutation = (
  { __typename?: 'Mutation' }
  & { createShippingCountry: (
    { __typename?: 'ShippingCountryPrice' }
    & Pick<ShippingCountryPrice, 'id' | 'price' | 'country_code' | 'country_name'>
  ) }
);

export type UpdateShippingCountryMutationVariables = Exact<{
  input: UpdateShippingCountryInput;
}>;


export type UpdateShippingCountryMutation = (
  { __typename?: 'Mutation' }
  & { updateShippingCountry: (
    { __typename?: 'ShippingCountryPrice' }
    & Pick<ShippingCountryPrice, 'id' | 'price' | 'country_code' | 'country_name'>
  ) }
);